Define _CRT_SECURE_NO_WARNINGS and silence remaining MSVC C4267
Define _CRT_SECURE_NO_WARNINGS globally on MSVC at the root
CMakeLists so that fopen/sprintf/strcpy etc. don't trip C4996. This
also lets us drop the per-call fopen_s shim in yolox_tflite.

Cast every remaining .size() (and .size() * sizeof(...)) being
passed to ailia API arguments declared as unsigned int / int:
- ailiaGetBlobData / ailiaSetInputBlobData (m2det, gpt-sovits,
  retinaface, g2p_en_model)
- ailiaPredict (clip)
- ailiaAudioGetFrameLen / ailiaAudioGetResampleLen / ailiaAudioResample
  (clap_utils, gpt-sovits)
- ailiaTokenizerDecode (sentence_transformers, fugumt-{ja,en}-{en,ja},
  multilingual-e5, t5_whisper_medical)

These are the C4267 size_t->unsigned int conversions that the MSVC
build still flagged.

+++ audio_processing/clap/clap_utils.cpp
@@ -25,7 +25,7 @@ static std::vector<float> get_mel_ailia(std::vector<float>& audio_data, const AU
     int frame_n;
     std::vector<float> mel;
     
-    status = ailiaAudioGetFrameLen(&frame_n, audio_data.size(), audio_cfg.window_size, audio_cfg.hop_size, center);
+    status = ailiaAudioGetFrameLen(&frame_n, (int)audio_data.size(), audio_cfg.window_size, audio_cfg.hop_size, center);
     if (status != AILIA_STATUS_SUCCESS) {
         PRINT_ERR("ailiaAudioGetFrameLen failed %d\n", status);
         return mel;
